Develop

Closes CAM-354, CAM-282, CAM-276, CAM-224, SF-200, CAM-199, CAM-250, SF-130, CAM-104, SF-196, SF-192, SF-135, SF-178, and SF-182

See merge request web-front-end/store-front!389

+++ components/Cam/ShippingAddress/includes/AddressBody.js
@@ -5,14 +5,22 @@ import { Row, Col } from 'react-bootstrap';
 import SVGComponent from '../../../common/SVGComponet';
 import { languageDefinations } from '../../../../utils/lang/';
 
-import { mergeCss } from '../../../../utils/cssUtil';
-const styles = mergeCss('components/Cam/ShippingAddress/address');
+import lang from '../../../../utils/language';
+
+import styles_en from '../address_en.styl';
+import styles_ar from '../address_ar.styl';
+
+
+const styles = lang === 'en' ? styles_en : styles_ar;
 
 const AddressBody = (props) => {
 
   const deleteAddr = (e) => {
-    props.deleteAddr(e.target.id || e.target.parentNode.id)
-  }
+    const confirmDelete = confirm("Are you sure you want to delete this address?");
+    if (confirmDelete) {
+      props.deleteAddr(e.target.id || e.target.parentNode.id)
+    }
+  };
 
   const editAddress = (e) => {
     props.editAddress(e.target.id || e.target.parentNode.id);
@@ -49,8 +57,8 @@ const AddressBody = (props) => {
                   </div>
                   <div className={styles['address-card-body']}>
                     <h5 className={`${styles['fontW600']} ${styles['m-0']} ${styles['mb-10']} ${styles['lgt-blue']}`}> {val.first_name + ' ' + val.last_name} </h5>
-                    <address className={`${styles['fs-12']} ${styles['thick-gry-clr']} ${styles['full-address-details']}`}>
-                      {val.address_line_1 + ', ' + val.address_line_2 + ', ' + val.city + ', ' + val.country}
+                    <address className={`${styles['fs-12']} ${styles['thick-gry-clr']} ${styles['full-address-details']}`} title={val.address_line_1 + ', ' + val.address_line_2 + ', ' + val.city + ', ' + val.country_name + ', ' + val.postal_code}>
+                      {val.address_line_1 + ', ' + val.address_line_2 + ', ' + val.city + ', ' + val.country_name + ', ' + val.postal_code}
                     </address>
                     <span className={`${styles['address-card-phone']} ${styles['fontW600']} ${styles['lgt-blue']}`}>
                       {val.mobile_country_code + ' ' + val.mobile_no}
@@ -78,7 +86,7 @@ const AddressBody = (props) => {
           })
         }
         <Col md={4} sm={12} xs={12}>
-          <div className={`${styles['address-card']} ${styles['address-card-new']} ${styles['p-20']}`} onClick={props.showAddAdrressForm}>
+          <div className={`${styles['address-card']} ${styles['address-card-new']} ${styles['p-20']}`} onClick={props.resetAddAdrressForm}>
             <div className={`${styles['flex-center']} ${styles['flex-wrap']}`}>
               <h5 className={`${styles['m-0']} ${styles['mb-10']} ${styles['thick-blue']} ${styles['fontW600']} ${styles['flex']}`}>
                 <SVGComponent clsName={`${styles['pls-icon']}`} src="icons/common-icon/plus-icon" />
@@ -86,7 +94,7 @@ const AddressBody = (props) => {
               </h5>
               <p className={`${styles['fs-12']} ${styles['thick-gry-clr']}`}>{DELIVERY_ADDR_PAGE.ADD_NEW_ADDR_TAG}</p>
               <p>
-                <button className={`${styles['fp-btn']} ${styles['fp-btn-default']} ${styles['text-uppercase']} ${styles['small-btn']}`}>
+                <button className={`${styles['fp-btn']} ${styles['fp-btn-default']} ${styles['small-btn']}`}>
                   {DELIVERY_ADDR_PAGE.ADD_NEW_BTN}
                 </button>
               </p>
@@ -102,6 +110,7 @@ AddressBody.propTypes = {
   deleteAddr: PropTypes.func.isRequired,
   editAddress: PropTypes.func.isRequired,
   makeDefaultAddress: PropTypes.func.isRequired,
+  resetAddAdrressForm: PropTypes.func.isRequired,
   data: PropTypes.array
 };
